Lmax cost normalized

+++ Java/src/problem_modelings/budget_constrained_lmax_optimization/algorithms/QABudgetConstrainedLmaxOptimizationModeling.java
@@ -124,7 +124,7 @@ public class QABudgetConstrainedLmaxOptimizationModeling extends BudgetConstrain
     @Override
     public Cost calculateCost(int currentReplicaNum) {
         int maxL = super.calculateMaxL(modelPlainOldData.tempControllerXSpinVariables);
-        int summationOfLMax = super.calculateDistanceToNearestControllerEnergy(modelPlainOldData.tempControllerXSpinVariables);
+        double summationOfLMax = super.calculateDistanceToNearestControllerEnergy(modelPlainOldData.tempControllerXSpinVariables);
 
         int reliabilityEnergy = Utils.getReliabilityEnergy(
                 modelPlainOldData.graph,
